+++ weave-js/src/components/PagePanelComponents/Home/Browse3.tsx
@@ -21,7 +21,6 @@ import {
 } from 'react-router-dom';
 
 import {MOON_200} from '../../../common/css/color.styles';
-import {useWindowSize} from '../../../common/hooks/useWindowSize';
 import {useWeaveContext} from '../../../context';
 import {useNodeValue} from '../../../react';
 import {URL_BROWSE3} from '../../../urls';
@@ -70,6 +69,7 @@ import {
 } from './Browse3/pages/wfInterface/naive';
 import {SideNav} from './Browse3/SideNav';
 import {useDrawerResize} from './useDrawerResize';
+import {useFlexDirection} from './useFlexDirection';
 
 LicenseInfo.setLicenseKey(
   '7684ecd9a2d817a3af28ae2a8682895aTz03NjEwMSxFPTE3MjgxNjc2MzEwMDAsUz1wcm8sTE09c3Vic2NyaXB0aW9uLEtWPTI='
@@ -325,16 +325,9 @@ const MainPeekingLayout: FC = () => {
     params.project!
   );
   const targetBase = baseRouter.projectUrl(params.entity!, params.project!);
-  const windowSize = useWindowSize();
-  const flexDirection = useMemo(() => {
-    if (windowSize.height > windowSize.width * 0.66) {
-      return 'column';
-    }
-    return 'row';
-  }, [windowSize.height, windowSize.width]);
+  const flexDirection = useFlexDirection();
 
-  const {handleMousedown, drawerWidth, drawerHeight} =
-    useDrawerResize(flexDirection);
+  const {handleMousedown, drawerWidth, drawerHeight} = useDrawerResize();
 
   return (
     <Box
